Revert "refactor(web): separate schemas from types and relax validation (#984)" (#997)
This reverts commit 782007f5b0df73288ae3d1da2d11757d2f71ce41.
