+++ packages/web/src/common/constants/web.constants.ts
@@ -128,9 +128,3 @@ export const ACCEPTED_TIMES = [
   "11:30 PM",
   "11:45 PM",
 ];
-
-export enum Recurrence_Selection {
-  NONE = "none",
-  WEEK = "week",
-  MONTH = "month",
-}
