+++ packages/web/src/common/utils/someday.util.ts
@@ -4,9 +4,11 @@ import { Schema_Event } from "@core/types/event.types";
 import { Categories_Event } from "@core/types/event.types";
 import { COLUMN_MONTH, COLUMN_WEEK } from "@web/common/constants/web.constants";
 import { ID_SOMEDAY_EVENT_ACTION_MENU } from "@web/common/constants/web.constants";
-import { Schema_SomedayEvent } from "@web/common/schemas/events/someday.event.schemas";
+import {
+  Schema_SomedayEvent,
+  Schema_SomedayEventsColumn,
+} from "@web/common/types/web.event.types";
 import { validateSomedayEvents } from "@web/common/validators/someday.event.validator";
-import { Schema_SomedayEventsColumn } from "@web/views/Calendar/components/Sidebar/sidebar.types";
 
 export const getSomedayEventCategory = (
   event: Schema_Event,
