+++ packages/web/src/ducks/events/sagas/someday.sagas.ts
@@ -1,7 +1,7 @@
 import { normalize } from "normalizr";
 import { call, put } from "redux-saga/effects";
 import { Schema_Event } from "@core/types/event.types";
-import { Schema_OptimisticEvent } from "@web/common/schemas/events/draft.event.schemas";
+import { Schema_OptimisticEvent } from "@web/common/types/web.event.types";
 import { handleError } from "@web/common/utils/event.util";
 import { setSomedayEventsOrder } from "@web/common/utils/someday.util";
 import { EventApi } from "@web/ducks/events/event.api";
