+++ packages/web/src/ducks/events/slices/event.slice.ts
@@ -4,7 +4,7 @@ import { createSlice } from "@reduxjs/toolkit";
 import { Schema_Event } from "@core/types/event.types";
 import { createAsyncSlice } from "@web/common/store/helpers";
 import { Response_HttpPaginatedSuccess } from "@web/common/types/api.types";
-import { Payload_NormalizedAsyncAction } from "@web/ducks/events/entity.types";
+import { Payload_NormalizedAsyncAction } from "@web/common/types/entity.types";
 import {
   Action_DeleteEvent,
   Action_EditEvent,
