+++ packages/web/src/ducks/events/slices/week.slice.ts
@@ -1,7 +1,7 @@
 import { PayloadAction } from "@reduxjs/toolkit";
 import { AsyncState, createAsyncSlice } from "@web/common/store/helpers";
 import { Response_HttpPaginatedSuccess } from "@web/common/types/api.types";
-import { Payload_NormalizedAsyncAction } from "@web/ducks/events/entity.types";
+import { Payload_NormalizedAsyncAction } from "@web/common/types/entity.types";
 import { Payload_GetEvents } from "@web/ducks/events/event.types";
 import { cancel, convert, insert } from "./someday.slice";
