+++ packages/web/src/views/Calendar/components/Sidebar/sidebar.types.ts
@@ -1,5 +1,4 @@
 import { Priorities } from "@core/constants/core.constants";
-import { Schema_Event } from "@core/types/event.types";
 
 export interface FutureEventsProps {
   shouldSetTopMargin?: boolean;
@@ -12,17 +11,4 @@ export interface PriorityFilter {
 
 export interface SectionProps {
   height?: string;
-}
-
-export interface Schema_SomedayEventsColumn {
-  columns: {
-    [key: string]: {
-      id: string;
-      eventIds: string[];
-    };
-  };
-  columnOrder: string[];
-  events: {
-    [key: string]: Schema_Event;
-  };
 }
